finalized bro_pcapAnalysis.sh

+++ bro_pcapAnalysis.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 #Script to automate pcapp analysis
+
+#Directory Paths for 
 PCAP_DIR="/home/is414/bro_is414/pcaps/*"
 BROLOGS_DIR="/home/is414/bro_is414/brologs/"
 
@@ -8,21 +10,35 @@ RED='\033[0;31m'
 GREEN='\033[0;32m'
 NC='\033[0m'
 
+echo "Welcome to bro_pcapAnalysis.sh"
+echo "I am a bot that will automate a few tasks for you"
+echo "I will look for pcap files located at $PCAP_DIR"
+echo -e "I will take each pcap file that I find and analyze it with the following command:\n /usr/local/bro/bin/bro -r [file]"
+echo -e "I will save the bro analysis in $BROLOGS_DIR\n"
 
-rm -rf $BROLOGS_DIR
-mkdir $BROLOGS_DIR
+read -p "Begin automation now? (Y/N)" -n 1 -r
 
-for file in $PCAP_DIR
-do
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+	echo -e "\nBegining Automation"
+	rm -rf $BROLOGS_DIR
+	mkdir $BROLOGS_DIR
 
-	#make directory for $file analysis
-	filename=$(basename -- "$file")
-	echo -e "${RED}Now analyzing $filename${NC}"
-	mkdir "${BROLOGS_DIR}${filename}"
-	cd /home/is414/bro_is414/brologs/"$filename"
+	for file in $PCAP_DIR
+	do
 
-	#analyze $file with bro
-	/usr/local/bro/bin/bro -r "$file"
+	        #make directory for $file analysis
+	        filename=$(basename -- "$file")
+	        echo -e "${RED}Now analyzing $filename${NC}"
+	        mkdir "${BROLOGS_DIR}${filename}"
+	        cd /home/is414/bro_is414/brologs/"$filename"
 
-	echo -e "${GREEN}Saved analysis in $(pwd)${NC}"
-done
+	        #analyze $file with bro
+	        /usr/local/bro/bin/bro -r "$file"
+
+	        echo -e "${GREEN}Saved analysis in $(pwd)${NC}"
+	done
+
+else
+	echo -e "\nExiting script without any actions"
+fi
